fix the ambigous overload error

+++ src/kde.cpp
@@ -136,7 +136,7 @@ std::vector<std::vector<double>> KernelDensityEstimation::compute_gaussian_kerne
 
     std::vector<double> weight;
     for (double width : bandwidth) {
-        weight.push_back(1.0 / pow(width, 2));
+        weight.push_back(1.0 / pow(width, 2.0));
     }
 
     std::vector<std::vector<double>> kernel_density_estimate(num, std::vector<double>(num));
@@ -157,7 +157,7 @@ std::vector<std::vector<double>> KernelDensityEstimation::compute_gaussian_kerne
         std::vector<std::vector<double>> &condition_variable, double bandwidth) {
     uint num = (uint) condition_variable.size();
     uint d = (uint) condition_variable[0].size();
-    double det = pow(bandwidth, d);
+    double det = pow(bandwidth, (double) d);
     double density = 1.0 / (pow(2 * CDC_PI, d / 2.0) * sqrt(det));
 
     std::vector<std::vector<double>> kernel_density_estimate(num, std::vector<double>(num));
@@ -165,7 +165,7 @@ std::vector<std::vector<double>> KernelDensityEstimation::compute_gaussian_kerne
         kernel_density_estimate[i][i] = density;
         for (uint j = 0; j < i; j++) {
             kernel_density_estimate[j][i] = exp(
-                    -0.5 * square_Euclidean_distance(condition_variable[i], condition_variable[j]) / pow(bandwidth, 2));
+                    -0.5 * square_Euclidean_distance(condition_variable[i], condition_variable[j]) / pow(bandwidth, 2.0));
             kernel_density_estimate[j][i] *= density;
             kernel_density_estimate[i][j] = kernel_density_estimate[j][i];
         }
